refactor: reorganize agent core files
