+++ apps/backend/src/agent-core/agent/events/agent-sse-log.ts
