+++ apps/backend/src/agent-core/agent/index.ts
@@ -1,14 +1,5 @@
 export {Agent} from './agent.js';
-export {agentPersistence} from './agent-persistence.js';
-export type {AgentSseLogReaderOptions} from './agent-sse-log.js';
-export {AgentSseLog} from './agent-sse-log.js';
-export {FileContentCache} from './file-content-cache.js';
-export {FileStatCheckResult, FileStatTracker} from './file-stat-tracker.js';
-export {TodoStore} from './todo-store.js';
-export type {
-  AgentEvent,
-  AgentEventStream,
-  AgentOptions,
-  AgentSnapshot,
-  AgentSnapshotOptions,
-} from './types.js';
+export type {AgentSseLogReaderOptions} from './events/agent-sse-log.js';
+export {agentPersistence} from './persistence/agent-persistence.js';
+export {FileStatCheckResult} from './state/file-stat-tracker.js';
+export type {AgentEventStream, AgentSnapshot} from './types.js';
