+++ apps/backend/src/agent-core/agent/state/file-content-cache.ts
