+++ apps/backend/src/agent-core/agent/state/file-stat-tracker.test.ts
