+++ apps/backend/src/agent-core/agent/state/todo-store.ts
