+++ apps/backend/src/agent-core/tool/types.ts
@@ -2,9 +2,9 @@ import type {SseSubAgentEvent} from '@omnicraft/sse-events';
 import type {ToolFailureData} from '@omnicraft/tool-schemas';
 import type {z} from 'zod';
 
-import type {FileContentCache} from '../agent/file-content-cache.js';
-import type {FileStatTracker} from '../agent/file-stat-tracker.js';
-import type {TodoStore} from '../agent/todo-store.js';
+import type {FileContentCache} from '../agent/state/file-content-cache.js';
+import type {FileStatTracker} from '../agent/state/file-stat-tracker.js';
+import type {TodoStore} from '../agent/state/todo-store.js';
 import type {LlmConfig} from '../llm-api/types.js';
 import type {SkillDefinition} from '../skill/skill-definition.js';
 import type {UserInteractionBridge} from '../user-interaction/index.js';
